backend/requirements.txt
asgiref==3.10.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
defusedxml==0.7.1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djoser==2.3.3
environs==14.5.0
idna==3.11
marshmallow==4.1.0
oauthlib==3.3.1
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.11.0
python-dotenv==1.2.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.6.3
